Closer to reality

+++ ParallelDfs/Data/Tree.cs
@@ -4,5 +4,5 @@ public class Tree
 {
     public Node? Root { get; set; }
 
-    public long NodesAmount { get; set; }
+    public int NodesAmount { get; set; }
 }
